project/LLM/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
langchain-community
bs4
pypdf
chromadb
faiss-cpu